+++ scripts/build-docker.sh
@@ -15,7 +15,7 @@ source "${mydir}/utils.sh"
 
 usage() {
   msg
-  msg "Usage: $0 [-h|--help] [-p|--package hoprd|cover-traffic-daemon] [-f|--force] [-n|--no-tags]"
+  msg "Usage: $0 [-h|--help] [-p|--package hoprd|hopr-cover-traffic-daemon] [-f|--force] [-n|--no-tags]"
   msg
   msg "Use -f to force a Docker build even though no environment can be found. This is useful for local testing. No additional docker tags will be applied though if no environment has been found which is in contrast to the normal execution of the script."
   msg
@@ -52,7 +52,7 @@ while (( "$#" )); do
   esac
 done
 
-if [ "${package:-}" != "hoprd" ] && [ "${package:-}" != "cover-traffic-daemon" ]; then
+if [ "${package:-}" != "hoprd" ] && [ "${package:-}" != "hopr-cover-traffic-daemon" ]; then
   msg "Error: unsupported package"
   usage
   exit 1