+++ packages/core/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - crawling: crawling is not block anymore, leads to faster crawling
 - heartbeat: every connection uses its own timer now
 - `randomInteger` function is now cryptographically safe
+- ECDSA signatures now use a more compact representation (64 instead 65 bytes)
 
 ### Fixed